@@ -7,7 +7,7 @@
     "npm": ">= 3.10.8"
   },
   "scripts": {
-    "build": "ng build -prod -aot && gulp",
+    "build": "ng build --prod --aot && gulp",
     "start": "ng serve",
     "start:https": "ng serve --ssl true --ssl-key ./dev_cert.key --ssl-cert ./dev_cert.crt",
     "test": "ng test --watch false",
@@ -15,23 +15,29 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/common": "^4.0.0",
-    "@angular/compiler": "^4.0.0",
-    "@angular/core": "^4.0.0",
-    "@angular/forms": "^4.0.0",
-    "@angular/http": "^4.0.0",
-    "@angular/platform-browser": "^4.0.0",
-    "@angular/platform-browser-dynamic": "^4.0.0",
-    "@angular/router": "^4.1.2",
+    "@angular/animations": "^6.1.10",
+    "@angular/common": "^6.1.10",
+    "@angular/compiler": "^6.1.10",
+    "@angular/core": "^6.1.10",
+    "@angular/forms": "^6.1.10",
+    "@angular/http": "^6.1.10",
+    "@angular/platform-browser": "^6.1.10",
+    "@angular/platform-browser-dynamic": "^6.1.10",
+    "@angular/platform-server": "^6.1.10",
+    "@angular/router": "^6.1.10",
     "core-js": "^2.4.1",
     "intl": "^1.2.5",
+    "ngx-paypal": "^5.0.0",
+    "ngx-swiper-wrapper": "^10.0.0",
     "node-sass": "^4.5.2",
-    "rxjs": "^5.1.0",
+    "rxjs": "^6.6.6",
+    "rxjs-compat": "^6.0.0-rc.0",
     "zone.js": "^0.8.4"
   },
   "devDependencies": {
-    "@angular/cli": "1.0.2",
-    "@angular/compiler-cli": "^4.0.0",
+    "@angular-devkit/build-angular": "~0.8.0",
+    "@angular/cli": "^6.2.9",
+    "@angular/compiler-cli": "6.1.10",
     "@types/jasmine": "2.5.38",
     "@types/node": "~6.0.60",
     "@types/sinon": "^2.2.1",
@@ -51,6 +57,6 @@
     "sinon": "^2.2.0",
     "ts-node": "~2.0.0",
     "tslint": "~4.5.0",
-    "typescript": "~2.2.0"
+    "typescript": "2.9.2"
   }
 }
