@@ -3,8 +3,7 @@ import { Product } from "app/models/product.model";
 import { ShoppingCart } from "app/models/shopping-cart.model";
 import { ProductsDataService } from "app/services/products.service";
 import { ShoppingCartService } from "app/services/shopping-cart.service";
-import { Observable } from "rxjs/Observable";
-import { Subscription } from "rxjs/Subscription";
+import { Observable ,  Subscription } from "rxjs";
 
 @Component({
   changeDetection: ChangeDetectionStrategy.OnPush,
