@@ -2,8 +2,7 @@ import { ChangeDetectionStrategy, Component, OnInit } from "@angular/core";
 import { Product } from "app/models/product.model";
 import { ProductsDataService } from "app/services/products.service";
 import { ShoppingCartService } from "app/services/shopping-cart.service";
-import { Observable } from "rxjs/Observable";
-import { Observer } from "rxjs/Observer";
+import { Observable ,  Observer } from "rxjs";
 
 @Component({
   changeDetection: ChangeDetectionStrategy.OnPush,
