@@ -1,7 +1,9 @@
+
 import { Injectable } from "@angular/core";
 import { Http } from "@angular/http";
+import { Observable } from "rxjs";
 import "rxjs/add/operator/map";
-import { Observable } from "rxjs/Observable";
+import {map} from "rxjs/operators";
 import { DeliveryOption } from "../models/delivery-option.model";
 import { CachcingServiceBase } from "./caching.service";
 
@@ -17,13 +19,13 @@ export class DeliveryOptionsDataService extends CachcingServiceBase {
     return this.cache<DeliveryOption[]>(() => this.deliveryOptions,
                                         (val: Observable<DeliveryOption[]>) => this.deliveryOptions = val,
                                         () => this.http
-                                                  .get("./assets/delivery-options.json")
-                                                  .map((response) => response.json()
+                                                  .get("./assets/delivery-options.json").pipe(
+                                                  map((response) => response.json()
                                                                              .map((item) => {
                                                                                 let model = new DeliveryOption();
                                                                                 model.updateFrom(item);
                                                                                 return model;
-                                                                              })));
+                                                                              }))));
 
   }
 }
